requirements.txt
Flask
Flask-Cors
gunicorn
python-dotenv
requests
stripe
